+++ packages/terra-form-select/src/_Frame.jsx
@@ -144,9 +144,16 @@ class Frame extends React.Component {
   }
 
   getDisplay() {
-    const { hasSearchChanged, searchValue } = this.state;
     const {
-      disabled, display, placeholder, variant,
+      hasSearchChanged,
+      searchValue,
+    } = this.state;
+
+    const {
+      disabled,
+      display,
+      placeholder,
+      variant,
     } = this.props;
 
     const inputAttrs = {
@@ -155,6 +162,8 @@ class Frame extends React.Component {
       ref: this.setInput,
       onChange: this.handleSearch,
       onMouseDown: this.handleInputMouseDown,
+      type: 'text',
+      'aria-owns': this.state.isOpen ? 'terra-select-menu' : undefined,
       'aria-label': 'search',
       'aria-disabled': disabled,
       className: cx('search-input', { 'is-hidden': Util.shouldHideSearch(this.props, this.state) }),
@@ -385,15 +394,21 @@ class Frame extends React.Component {
       customProps.className,
     ]);
 
+    let role;
+
+    if (!disabled) {
+      role = 'combobox';
+    }
+
     return (
       <div
         {...customProps}
-        role={!disabled ? 'combobox' : undefined}
-        aria-controls={!disabled && this.state.isOpen ? 'terra-select-dropdown' : undefined}
+        role={role}
+        aria-controls={!disabled && this.state.isOpen ? 'terra-select-menu' : undefined}
         aria-disabled={!!disabled}
         aria-expanded={!!disabled && !!this.state.isOpen}
-        aria-haspopup={!disabled ? 'true' : undefined}
-        aria-owns={this.state.isOpen ? 'terra-select-dropdown' : undefined}
+        aria-haspopup={!disabled ? 'listbox' : undefined}
+        aria-owns={this.state.isOpen ? 'terra-select-menu' : undefined}
         className={selectClasses}
         onBlur={this.handleBlur}
         onFocus={this.handleFocus}
@@ -409,7 +424,7 @@ class Frame extends React.Component {
         <div className={cx('toggle')} onMouseDown={this.toggleDropdown}>
           <span className={cx('arrow-icon')} />
         </div>
-        <span className={cx('visually-hidden-component')} ref={this.visuallyHiddenComponent} aria-live="polite" aria-relevant="additions text" aria-atomic="true" />
+        <span className={cx('visually-hidden-component')} ref={this.visuallyHiddenComponent} aria-live="assertive" aria-relevant="additions text" aria-atomic="true" />
         {this.state.isOpen
           && (
           <Dropdown
@@ -429,6 +444,7 @@ class Frame extends React.Component {
                  onDeselect,
                  optionFilter,
                  noResultContent,
+                 focusRegion: variant === Variants.DEFAULT ? this.select : this.input,
                  visuallyHiddenComponent: this.visuallyHiddenComponent,
                  onSelect: this.handleSelect,
                  onRequestClose: this.closeDropdown,