+++ packages/terra-form-select/tests/jest/Menu.test.jsx
@@ -56,4 +56,40 @@ describe('Menu', () => {
     expect(wrapper).toMatchSnapshot();
     expect(liveRegion.current.innerText).toEqual('No matching results for "asdf"');
   });
+
+  it('should update the search input aria-descendent attribute and visually hidden text when active element changes', () => {
+    const liveRegion = { current: document.createElement('div') };
+    const input = document.createElement('input');
+    input.setAttribute('type', 'text');
+
+    const menu = (
+      <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} focusRegion={input} variant="default" searchValue="display">
+        <Option value="value" display="display" />
+        <Option value="value_two" display="display" />
+      </Menu>
+    );
+
+    const wrapper = mount(menu, intlContexts.shallowContext);
+    wrapper.setState({ active: 'value_two' });
+    wrapper.instance().updateCurrentActiveScreenReader();
+    expect(input.getAttribute('aria-activedescendant')).toEqual('terra-select-option-value_two');
+
+    expect(liveRegion.current.innerText).toEqual('display');
+  });
+
+  it('should update the visually hidden text with Selected text when a selected option is active', () => {
+    const liveRegion = { current: document.createElement('div') };
+
+    const menu = (
+      <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} variant="default" searchValue="display" value="value_two">
+        <Option value="value" display="display" />
+        <Option value="value_two" display="display" />
+      </Menu>
+    );
+
+    const wrapper = mount(menu, intlContexts.shallowContext);
+    wrapper.setState({ active: 'value_two' });
+    wrapper.instance().updateCurrentActiveScreenReader();
+    expect(liveRegion.current.innerText).toEqual('display Selected');
+  });
 });